refactor: merge services
