+++ src/main/java/biblivre/administration/permissions/PermissionBO.java
@@ -20,6 +20,10 @@
 package biblivre.administration.permissions;
 
 import biblivre.circulation.user.UserDTO;
+import biblivre.core.SchemaThreadLocal;
+import biblivre.core.auth.AuthorizationPoints;
+import biblivre.core.exceptions.AuthorizationException;
+import biblivre.login.LoginDTO;
 import java.util.Collection;
 import java.util.List;
 
@@ -45,5 +49,26 @@ public class PermissionBO {
 
     public void setPermissionDAO(PermissionDAO permissionDAO) {
         this.permissionDAO = permissionDAO;
+    }
+
+    public void authorize(AuthorizationPoints atps, String module, String action) {
+        if (atps == null) {
+            atps = AuthorizationPoints.getNotLoggedInstance();
+        }
+
+        if (!atps.isAllowed(module, action)) {
+            throw new AuthorizationException();
+        }
+    }
+
+    public AuthorizationPoints getUserAuthorizationPoints(LoginDTO login) {
+        Collection<String> permissions = null;
+
+        if (!SchemaThreadLocal.isGlobalSchema()) {
+            permissions = this.permissionDAO.getByLoginId(login.getId());
+        }
+
+        return new AuthorizationPoints(
+                SchemaThreadLocal.get(), true, login.isEmployee(), permissions);
     }
 }
