--- src/main/java/biblivre/core/auth/AuthorizationDAO.java
@@ -1,9 +1,0 @@
-package biblivre.core.auth;
-
-import biblivre.login.LoginDTO;
-import java.util.Map;
-
-public interface AuthorizationDAO {
-
-    Map<String, Boolean> getUserPermissions(LoginDTO user);
-}
